prepare to query for traffic times and change notifications
