+++ server/db/controllers/travelController.js
@@ -16,13 +16,14 @@ const initiateTravel = function(event, initialEstimate) {
       // trafficEstimate: Sequelize.INTEGER,
       // mapsUrl: Sequelize.STRING,
       notificationTime: new Date(Date.parse(event.startDateTime) - (initialEstimate + 300000)),
-      userId: event.userId
       // notify 5 minutes before you need to leave
+      userId: event.userId
     }})
   .spread(function(travel, created) {
     console.log(created, ': travel was created');
       // schedule notification to be sent based on initial travel time, see workers/mapsWorker.js
     agenda.schedule(travel.dataValues.notificationTime, 'send notification', travel.dataValues);
+    // instead of scheduling send notification, this should schedule 'query maps for traffic' job, which will then schedule a notifcation;
   });
 }
