+++ server/utility/googleMaps.js
@@ -8,8 +8,6 @@ const url = 'https://maps.googleapis.com/maps/api/distancematrix/json';
 const origins = 'Hack Reactor, 944 Market St, San Francisco, CA 94102';
 
 const getInitialTravelTime = function(event) {
-  // Event.retrieveEvent(eventId)
-  // .then((event) => {
     var options = {
       url,
       qs: {
@@ -32,14 +30,46 @@ const getInitialTravelTime = function(event) {
         var distance = body.rows[0].elements[0].distance;
         var duration = body.rows[0].elements[0].duration;
 
-        // add data to Travel table, see /db/controlelrs/travelController
+        // add data to Travel table, see /db/controllers/travelController
         Travel.initiateTravel(event, (duration.value * 1000)); // convert seconds to milliseconds
       }
     });
 };
 
+const getTrafficTime = function(eventId) {
+  Event.retrieveEvent(eventId)
+  .then((event) => {
+    var options = {
+      url,
+      qs: {
+        key: process.env.GOOGLE_MAPS_API_KEY,
+        origins: origins, 
+        // origins needs to be changed to the user's current location or previous location
+        destinations: event.dataValues.location,
+        // can probably store destinations in job? 
+        mode: 'driving',
+        departure_time: 'now', 
+        units: 'imperial',
+        traffic_model: 'best_guess'
+      }
+    };
+    request(options, function (error, response, body) {
+      console.log(response.statusCode);
+      if (!error && response.statusCode == 200) {
+        body = JSON.parse(body);
+        console.log(body.rows[0].elements[0]);
+        var distance = body.rows[0].elements[0].distance;
+        var duration = body.rows[0].elements[0].duration;
+        var traffic = duration.value * 1000; // convert seconds to milliseconds
+      }
+    })
+  })
+};
+
+
 module.exports = {
-  getInitialTravelTime
+  getInitialTravelTime,
+  getTrafficTime
 }
 
 /* GAME PLAN
